requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
class-registry==2.1.2
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-simple-history==3.2.0
djangorestframework==3.14.0
djangorestframework-filter==0.2.2
djangorestframework-simplejwt==5.2.2
filters==1.3.2
filters-django==1.0.5
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
